+++ WorldEditCommands/WorldEditCommands.cs
@@ -7,7 +7,7 @@ namespace WorldEditCommands;
 public class WorldEditCommands : BaseUnityPlugin {
   public const string GUID = "world_edit_commands";
   public const string NAME = "World Edit Commands";
-  public const string VERSION = "1.38";
+  public const string VERSION = "1.39";
   public void Awake() {
     new Harmony(GUID).PatchAll();
   }